requirements.txt
Django==2.2.7
djangorestframework==3.10.3
django-cities-light==3.5.0
django-extensions==2.2.5
django-rest-swagger==2.2.0
drf-firebase-authentication==1.0.1
Pillow==6.2.1
drf-yasg==1.17.0